requirements.txt
langchain
python-dotenv
langchain-core
langchain-community
bs4
faiss-cpu
langchain_huggingface
langchain-groq
python-dotenv
flask
flask-cors
sentence-transformers
fastapi
uvicorn
keras